backend/requirements.txt
# RemoteLED Backend API Requirements
# Python 3.10+

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
asyncpg==0.29.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Cryptography (for ECDSA signing)
ecdsa==0.18.0
cryptography==41.0.7

# Payments

# Authentication
passlib[bcrypt]==1.7.4
firebase-admin==6.3.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0

# Date/Time
python-dateutil==2.8.2

# Bluetooth LE
bleak==0.22.3

# Email
aiosmtplib==3.0.1
jinja2==3.1.2

# Testing (optional)
pytest==7.4.3
httpx==0.25.2
